Read on chapters 1 through 3
This is really exciting. The simple neural networks we have created so
far are indeed basic but just the structure and idea behind them is
amazing. For now, the networks are not learning, infact, all they know
is hard coded in the source files, in chapter 4, we will begin making
them learn by adjusting their weights until a good configuration is
found that predicts better. Amazingly, persisting the network and all
it knows is a matter of serializing the weights and loading them when
starting up the network.